Accel

+++ install.py
@@ -16,9 +16,27 @@ elif pf == "Darwin":
     print("Now installing...")
     #os.system("sh ./setup/setup_macOS")
     os.system("python3 "+str(os.path.abspath("./"))+"/setup/setup.py")
-    os.system("sh "+str(os.path.abspath("./"))+"/install/install")
-    os.system("sudo ln -si $PWD /opt/plantfem")
-    os.system("sudo ln -si $PWD/plantfem /usr/local/bin/plantfem")
+    
+    if os.path.exists("/opt/plantfem/inc/obj.o"):
+        print("plantFEM is already built.")
+    else:
+        os.system("sh "+str(os.path.abspath("./"))+"/install/install")
+    
+    if os.path.exists("/opt/plantfem"):
+        print("plantFEM is already installed.")
+    else:
+        os.system("sudo ln -si $PWD /opt/plantfem")
+    
+    
+    if os.path.exists("/usr/local/bin/plantfem"):
+        print("plantFEM is already installed.")
+    else:
+        os.system("sudo ln -si $PWD/plantfem /usr/local/bin/plantfem")
+    
+    if os.path.exists("/usr/local/bin/plantfem/soja"):
+        print("soja (package manager of plantFEM) is already installed.")
+    else:
+        os.system("sudo ln -si "+str(os.path.abspath("./"))+"/bin/soja.sh /usr/local/bin/soja")
     #print("[Next!]  Please type")
     #print("export PATH=$PATH:$PWD")
     #print("export PATH=$PATH:$PWD/bin")
@@ -31,9 +49,20 @@ elif pf == "Linux":
     print("Now installing...")
     #os.system("sh ./setup/setup")
     os.system("python3 "+str(os.path.abspath("./"))+"/setup/setup.py")
-    os.system("sh "+str(os.path.abspath("./"))+"/install/install")
-    os.system("sudo ln -si "+str(os.path.abspath("./"))+"/plantfem /usr/local/bin/plantfem")
-    os.system("sudo ln -si "+str(os.path.abspath("./"))+"/bin/soja.sh /usr/local/bin/soja")
+    if os.path.exists("/opt/plantfem/inc/obj.o"):
+        print("plantFEM is already built.")
+    else:
+        os.system("sh "+str(os.path.abspath("./"))+"/install/install")
+        
+    if os.path.exists("/usr/local/bin/plantfem"):
+        print("plantFEM is already installed.")
+    else:
+        os.system("sudo ln -si "+str(os.path.abspath("./"))+"/plantfem /usr/local/bin/plantfem")
+    
+    if os.path.exists("/usr/local/bin/soja"):
+        print("soja (package manager of plantFEM) is already installed.")
+    else:
+        os.system("sudo ln -si "+str(os.path.abspath("./"))+"/bin/soja.sh /usr/local/bin/soja")
     #os.system("sudo ln -si "+str(os.path.abspath("./"))+"/bin/plantfem_run.py /usr/local/bin/plantfem_run.py")
     #os.system("sudo ln -si "+str(os.path.abspath("./"))+"/Interactive/plantfem_run /usr/local/bin/plantfem_run")
     os.system("sudo ln -si $PWD /opt/plantfem")
